+++ src/commands/osu_maps/MapParse.ts
@@ -82,6 +82,9 @@ export class MapParse extends OsuCommand {
 
         //get map id via mapset if not in the given URL
         await this.setParamMapGone(mapTemp);
+        if (!this.params.mapid) {
+            this.params.mapid = this.setParam(this.params.mapid, helper.argflags.beatmap, 'number', { number_isInt: true });
+        }
     }
     async setParamsInteract() {
         const interaction = this.input.interaction as Discord.ChatInputCommandInteraction;