+++ src/commands/osu_scores/MapLeaderboard.ts
@@ -59,6 +59,9 @@ export class MapLeaderboard extends OsuCommand {
                 }
             }
         }
+        if (!this.params.mapid) {
+            this.params.mapid = this.setParam(this.params.mapid, helper.argflags.beatmap, 'number', { number_isInt: true });
+        }
     }
     async setParamsInteract() {
         const interaction = this.input.interaction as Discord.ChatInputCommandInteraction;