+++ src/commands/osu_scores/ScoreListCommand.ts
@@ -105,19 +105,7 @@ export class ScoreListCommand extends OsuCommand {
         this.params.filteredMapper = this.setParam(this.params.filteredMapper, helper.argflags.filterCreator, 'string', { string_isMultiple: true });
         this.params.filterArtist = this.setParam(this.params.filterArtist, helper.argflags.filterArtist, 'string', { string_isMultiple: true });
         this.params.filterDifficulty = this.setParam(this.params.filterDifficulty, helper.argflags.filterVersion, 'string', { string_isMultiple: true });
-        this.params.filterRank = this.setParamBoolList(this.params.filterRank,
-            { set: 'xh', flags: ['-xh'] },
-            { set: 'ssh', flags: ['-ssh'] },
-            { set: 'x', flags: ['-x'] },
-            { set: 'ss', flags: ['-ss'] },
-            { set: 'sh', flags: ['-sh'] },
-            { set: 's', flags: ['-s'] },
-            { set: 'a', flags: ['-a'] },
-            { set: 'b', flags: ['-b'] },
-            { set: 'c', flags: ['-c'] },
-            { set: 'd', flags: ['-d'] },
-            { set: 'f', flags: ['-f'] },
-        );
+        this.params.filterRank = this.setParam(this.params.filterRank, ['-rank'], 'string', { string_isMultiple: false }) as osuapi.types_v2.Rank;
         this.modParams();
         await this.paramsMsgExtra();
 
@@ -538,11 +526,7 @@ export class ScoreListCommand extends OsuCommand {
                 .setStyle(helper.buttons.type.current)
                 .setEmoji(helper.buttons.label.extras.user),
         );
-        try {
-            await this.getScores();
-        } catch (e) {
-            await this.sendError(helper.errors.scores.generic(this.params.user));
-        }
+        await this.getScores();
 
         if (this.params.parseScore) {
             const user = this.osudata.username;