+++ changelog.md
@@ -2,9 +2,21 @@
 
 # changelog
 
+
 ## [WIP] - xxxx-xx-xx
 
 [commit](https://github.com/sbrstrkkdwmdr/ssob)</br>
+
+## [5.3.2] - 2025-08-26
+
+[commit](https://github.com/sbrstrkkdwmdr/ssob/commit/385e6c253860790f7e1444520dee8d28aaa7a4be)</br>
+
+### Fixed
+
+-   score list sending multiple errors if empty
+-   "F if pass" on score list commands
+-   missing prefix in command examples
+-   commands not accepting mapid via flags
 
 ## [5.3.1] - 2025-08-08